+++ src/lib/init-hooks/core-hooks/react/dynamicComponents/beforeEntryLoader.js
@@ -11,18 +11,7 @@
 export default function beforeEntryLoader() {
 
   // THIS IS A GENERATED FILE, PLEASE DO NOT MODIFY
-  const Loadable = require('react-loadable');
-  const React = require('react');
 
-  const containers = {};
-  const modules = {};
-  const decorators = {};
-
-  if (!window.__boring_internals) {
-    window.__boring_internals = {
-      wutsthis: 'DO NOT LOOK HERE OR YOU ARE FIRED',
-    };
-  }
 
   const internals = window.__boring_internals;
